+++ src/main/java/hungteen/htlib/common/network/DummyEntityPacket.java
@@ -21,9 +21,13 @@ import java.util.function.Supplier;
 public class DummyEntityPacket {
 
     private final Operation operation;
-    private final int entityID;
+    private int entityID;
     private ResourceLocation entityType;
     private CompoundTag entityNBT;
+
+    public DummyEntityPacket() {
+        this.operation = Operation.CLEAR;
+    }
 
     public DummyEntityPacket(DummyEntity dummyEntity, CompoundTag nbt) {
         this(Operation.UPDATE, dummyEntity);
@@ -33,7 +37,7 @@ public class DummyEntityPacket {
     public DummyEntityPacket(Operation operation, DummyEntity dummyEntity) {
         this.operation = operation;
         this.entityID = dummyEntity.getEntityID();
-        if(this.operation == Operation.CREATE){
+        if (this.operation == Operation.CREATE) {
             this.entityType = dummyEntity.getEntityType().getLocation();
             this.entityNBT = dummyEntity.save(new CompoundTag());
         }
@@ -41,46 +45,52 @@ public class DummyEntityPacket {
 
     public DummyEntityPacket(FriendlyByteBuf buffer) {
         this.operation = Operation.values()[buffer.readInt()];
-        this.entityID = buffer.readInt();
-        if(this.operation == Operation.CREATE){
-            this.entityType = new ResourceLocation(buffer.readUtf());
-            this.entityNBT = buffer.readNbt();
-        } else if(this.operation == Operation.UPDATE){
-            this.entityNBT = buffer.readNbt();
+        if (this.operation != Operation.CLEAR) {
+            this.entityID = buffer.readInt();
+            if (this.operation == Operation.CREATE) {
+                this.entityType = new ResourceLocation(buffer.readUtf());
+                this.entityNBT = buffer.readNbt();
+            } else if (this.operation == Operation.UPDATE) {
+                this.entityNBT = buffer.readNbt();
+            }
         }
     }
 
     public void encode(FriendlyByteBuf buffer) {
         buffer.writeInt(this.operation.ordinal());
-        buffer.writeInt(this.entityID);
-        if(this.operation == Operation.CREATE){
-            buffer.writeUtf(this.entityType.toString());
-            buffer.writeNbt(this.entityNBT);
-        } else if(this.operation == Operation.UPDATE){
-            buffer.writeNbt(this.entityNBT);
+        if (this.operation != Operation.CLEAR) {
+            buffer.writeInt(this.entityID);
+            if (this.operation == Operation.CREATE) {
+                buffer.writeUtf(this.entityType.toString());
+                buffer.writeNbt(this.entityNBT);
+            } else if (this.operation == Operation.UPDATE) {
+                buffer.writeNbt(this.entityNBT);
+            }
         }
     }
 
     public static class Handler {
 
         public static void onMessage(DummyEntityPacket message, Supplier<NetworkEvent.Context> ctx) {
-            ctx.get().enqueueWork(()->{
-                if(message.operation == Operation.CREATE){
+            ctx.get().enqueueWork(() -> {
+                if (message.operation == Operation.CREATE) {
                     HTDummyEntities.getEntityType(message.entityType).ifPresent(type -> {
                         Optional<Level> world = LogicalSidedProvider.CLIENTWORLD.get(ctx.get().getDirection().getReceptionSide());
                         DummyEntity dummyEntity = world.map(w -> type.create(w, message.entityNBT)).orElse(null);
                         if (dummyEntity != null) {
                             HTLib.PROXY.addDummyEntity(dummyEntity);
-                        } else{
+                        } else {
                             HTLib.getLogger().error("Fail to sync Dummy Entity on client side");
                         }
                     });
-                } else if(message.operation == Operation.REMOVE){
+                } else if (message.operation == Operation.REMOVE) {
                     HTLib.PROXY.removeDummyEntity(message.entityID);
-                } else if(message.operation == Operation.UPDATE){
+                } else if (message.operation == Operation.UPDATE) {
                     HTLib.PROXY.getDummyEntity(message.entityID).ifPresent(dummyEntity -> {
                         dummyEntity.load(message.entityNBT);
                     });
+                } else if (message.operation == Operation.CLEAR) {
+                    HTLib.PROXY.clearDummyEntities();
                 }
             });
             ctx.get().setPacketHandled(true);
@@ -90,8 +100,14 @@ public class DummyEntityPacket {
     public enum Operation {
 
         CREATE,
+
         REMOVE,
-        UPDATE
+
+        UPDATE,
+
+        CLEAR,
+
+        ;
 
     }