+++ src/main/java/hungteen/htlib/common/network/NetworkHandler.java
@@ -37,6 +37,10 @@ public class NetworkHandler {
         CHANNEL.sendToServer(msg);
     }
 
+    public static <MSG> void sendToClient(MSG msg){
+        CHANNEL.send(PacketDistributor.ALL.noArg(), msg);
+    }
+
     public static <MSG> void sendToClient(ServerPlayer serverPlayer, MSG msg){
         CHANNEL.send(PacketDistributor.PLAYER.with(() -> serverPlayer), msg);
     }