+++ docs/version/en_us/changelog.md
@@ -7,6 +7,11 @@
 * Add more helper methods.
 #### Other Changes
 * Temporarily cancelled compat with KubeJS.
+### 1.19.2-0.9.1
+#### New Features
+* More command to control dummy entity (include remove nearby and all).
+* improve network sync of dummy entity.
+* Fix dummy entities exist across different save.
 ### 1.19.2-0.9.0
 #### New Features
 * KubeJs support.